+++ x/epochs/keeper/keeper_test.go
@@ -33,8 +33,8 @@ func TestUpsertEpochInfo_HappyPath(t *testing.T) {
 
 	require.Len(t, allEpochs, 4)
 	// Epochs are ordered in alphabetical order
-	require.Equal(t, "15 min", allEpochs[0].Identifier)
-	require.Equal(t, "30 min", allEpochs[1].Identifier)
+	require.Equal(t, "30 min", allEpochs[0].Identifier)
+	require.Equal(t, "day", allEpochs[1].Identifier)
 	require.Equal(t, "monthly", allEpochs[2].Identifier)
 	require.Equal(t, "week", allEpochs[3].Identifier)
 }