+++ x/inflation/keeper/inflation_test.go
@@ -14,58 +14,105 @@ import (
 	"github.com/NibiruChain/nibiru/app"
 	"github.com/NibiruChain/nibiru/x/common/denoms"
 	"github.com/NibiruChain/nibiru/x/common/testutil/testapp"
-	types "github.com/NibiruChain/nibiru/x/inflation/types"
+	"github.com/NibiruChain/nibiru/x/inflation/types"
+	sudotypes "github.com/NibiruChain/nibiru/x/sudo/types"
 )
+
+func init() {
+	testapp.EnsureNibiruPrefix()
+}
 
 func TestMintAndAllocateInflation(t *testing.T) {
 	testCases := []struct {
-		name                    string
-		mintCoin                sdk.Coin
-		expStakingRewardAmt     sdk.Coin
-		expStrategicReservesAmt sdk.Coin
-		expCommunityPoolAmt     sdk.DecCoins
+		name                             string
+		coinsToMint                      sdk.Coin
+		expectedStakingAmt               sdk.Coin
+		expectedStrategicAmt             sdk.Coin
+		expectedCommunityAmt             sdk.Coin
+		expectedStakingRewardsBalance    sdk.Coin
+		expectedStrategicReservesBalance sdk.Coin
+		expectedCommunityPoolBalance     sdk.DecCoins
+		rootAccount                      string
 	}{
 		{
-			"pass",
-			sdk.NewCoin(denoms.NIBI, sdk.NewInt(1_000_000)),
-			sdk.NewCoin(denoms.NIBI, sdk.NewInt(278_000)),
-			sdk.NewCoin(denoms.NIBI, sdk.NewInt(100_000)),
-			sdk.NewDecCoins(sdk.NewDecCoin(denoms.NIBI, sdk.NewInt(622_000))),
+			name:                             "pass",
+			coinsToMint:                      sdk.NewCoin(denoms.NIBI, sdk.NewInt(1_000_000)),
+			expectedStakingAmt:               sdk.NewCoin(denoms.NIBI, sdk.NewInt(278_000)),
+			expectedStrategicAmt:             sdk.NewCoin(denoms.NIBI, sdk.NewInt(100_000)),
+			expectedCommunityAmt:             sdk.NewCoin(denoms.NIBI, sdk.NewInt(622_000)),
+			expectedStakingRewardsBalance:    sdk.NewCoin(denoms.NIBI, sdk.NewInt(278_000)),
+			expectedStrategicReservesBalance: sdk.NewCoin(denoms.NIBI, sdk.NewInt(100_000)),
+			expectedCommunityPoolBalance:     sdk.NewDecCoins(sdk.NewDecCoin(denoms.NIBI, sdk.NewInt(622_000))),
+			rootAccount:                      "nibi1qyqf35fkhn73hjr70442fctpq8prpqr9ysj9sn",
 		},
 		{
-			"pass - no coins minted ",
-			sdk.NewCoin(denoms.NIBI, sdk.ZeroInt()),
-			sdk.NewCoin(denoms.NIBI, sdk.ZeroInt()),
-			sdk.NewCoin(denoms.NIBI, sdk.ZeroInt()),
-			sdk.DecCoins(nil),
+			name:                             "pass - no coins minted ",
+			coinsToMint:                      sdk.NewCoin(denoms.NIBI, sdk.ZeroInt()),
+			expectedStakingAmt:               sdk.Coin{},
+			expectedStrategicAmt:             sdk.Coin{},
+			expectedCommunityAmt:             sdk.Coin{},
+			expectedStakingRewardsBalance:    sdk.NewCoin(denoms.NIBI, sdk.ZeroInt()),
+			expectedStrategicReservesBalance: sdk.NewCoin(denoms.NIBI, sdk.ZeroInt()),
+			expectedCommunityPoolBalance:     nil,
+			rootAccount:                      "nibi1qyqf35fkhn73hjr70442fctpq8prpqr9ysj9sn",
+		},
+		{
+			name:                             "pass - no root account",
+			coinsToMint:                      sdk.NewCoin(denoms.NIBI, sdk.NewInt(1_000_000)),
+			expectedStakingAmt:               sdk.NewCoin(denoms.NIBI, sdk.NewInt(278_000)),
+			expectedStrategicAmt:             sdk.NewCoin(denoms.NIBI, sdk.NewInt(100_000)),
+			expectedCommunityAmt:             sdk.NewCoin(denoms.NIBI, sdk.NewInt(622_000)),
+			expectedStakingRewardsBalance:    sdk.NewCoin(denoms.NIBI, sdk.NewInt(278_000)),
+			expectedStrategicReservesBalance: sdk.NewCoin(denoms.NIBI, sdk.NewInt(100_000)),
+			expectedCommunityPoolBalance:     sdk.NewDecCoins(sdk.NewDecCoin(denoms.NIBI, sdk.NewInt(622_000))),
+			rootAccount:                      "",
 		},
 	}
 	for _, tc := range testCases {
 		t.Run(fmt.Sprintf("Case %s", tc.name), func(t *testing.T) {
 			nibiruApp, ctx := testapp.NewNibiruTestAppAndContext()
 
-			_, _, _, err := nibiruApp.InflationKeeper.MintAndAllocateInflation(ctx, tc.mintCoin, types.DefaultParams())
+			if tc.rootAccount != "" {
+				t.Logf("setting root account to %s", tc.rootAccount)
+				nibiruApp.SudoKeeper.Sudoers.Set(ctx, sudotypes.Sudoers{
+					Root:      sdk.MustAccAddressFromBech32(tc.rootAccount).String(),
+					Contracts: []string{},
+				})
+			}
+
+			staking, strategic, community, err := nibiruApp.InflationKeeper.MintAndAllocateInflation(ctx, tc.coinsToMint, types.DefaultParams())
+			require.NoError(t, err)
+			assert.Equal(t, tc.expectedStakingAmt, staking)
+			assert.Equal(t, tc.expectedStrategicAmt, strategic)
+			assert.Equal(t, tc.expectedCommunityAmt, community)
 
 			// Get balances
-			balanceInflationModule := nibiruApp.BankKeeper.GetBalance(
-				ctx,
-				nibiruApp.AccountKeeper.GetModuleAddress(types.ModuleName),
-				denoms.NIBI,
-			)
+			var balanceStrategicReserve sdk.Coin
+			if tc.rootAccount != "" {
+				strategicAccount, err := nibiruApp.SudoKeeper.GetRoot(ctx)
+				require.NoError(t, err)
+				balanceStrategicReserve = nibiruApp.BankKeeper.GetBalance(
+					ctx,
+					strategicAccount,
+					denoms.NIBI,
+				)
+			} else {
+				// if no root account is specified, then the strategic reserve remains in the x/inflation module account
+				balanceStrategicReserve = nibiruApp.BankKeeper.GetBalance(ctx, nibiruApp.AccountKeeper.GetModuleAddress(types.ModuleName), denoms.NIBI)
+			}
 
-			feeCollector := nibiruApp.AccountKeeper.GetModuleAddress(authtypes.FeeCollectorName)
 			balanceStakingRewards := nibiruApp.BankKeeper.GetBalance(
 				ctx,
-				feeCollector,
+				nibiruApp.AccountKeeper.GetModuleAddress(authtypes.FeeCollectorName),
 				denoms.NIBI,
 			)
 
 			balanceCommunityPool := nibiruApp.DistrKeeper.GetFeePoolCommunityCoins(ctx)
 
 			require.NoError(t, err, tc.name)
-			assert.Equal(t, tc.expStakingRewardAmt, balanceStakingRewards)
-			assert.Equal(t, tc.expStrategicReservesAmt, balanceInflationModule)
-			assert.Equal(t, tc.expCommunityPoolAmt, balanceCommunityPool)
+			assert.Equal(t, tc.expectedStakingRewardsBalance, balanceStakingRewards)
+			assert.Equal(t, tc.expectedStrategicReservesBalance, balanceStrategicReserve)
+			assert.Equal(t, tc.expectedCommunityPoolBalance, balanceCommunityPool)
 		})
 	}
 }